+++ polus-feature-extraction-plugin/VERSION
@@ -1,1 +1,1 @@
-0.7.1
+0.7.2
